Create README.md
Hi, I’m @ChristineEconomides. I would like to use the Chicago Style Manual of Style 17th Edition (author-date) with one small modification: I do not want automatic parentheses in the citations because I prefer to use active verbs when possible, and it looks dumb to start a sentence with something in parentheses. Can someone help me how to create a modified style with only this change?